pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.4",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "jsonschema>=4.26.0",
    "jupyter-client>=8.8.0",
    "matplotlib>=3.10.9",
    "omegaconf>=2.3.0",
    "openai>=2.0.0",
    "pandas>=3.0.2",
    "peft>=0.19.1",
    "python-dotenv>=1.0.0",
    "pyzmq>=27.1.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "swebench",
    "tensorboard>=2.20.0",
    "torch>=2.6.0",
    "tqdm>=4.67.3",
    "transformers==4.57.1",
    "tree-sitter==0.21.3",
]

[tool.uv]
environments = ["sys_platform == 'linux'"]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }
swebench = { git = "https://github.com/SWE-rebench/SWE-bench-fork.git" }
